move tag commands out of interfaces
