+++ core/commands/__init__.py
@@ -20,3 +20,4 @@ from .reset import *
 from .custom import *
 from .flow import *
 from .cherry_pick import *
+from .tag import *
